Compare results with sortables if at least one has a value
We should not revert to by-id sorting if one of them lack the sorting
parameter; rather revert to the one which has the value as being higher
than the one which does not have it. This fixes a failing test (whose
failure was masked by a bad string comparison algorithm).